// backend/src/models/formats/progressFormat.d.ts
interface ReqProgressFormat {
  userId: string;
  progress1: boolean;
  progress2: boolean;
  progress3: boolean;
  progress4: boolean;
  progress5: boolean;
  progress6: boolean;
  progress7: boolean;
  progress8: boolean;
  progress9: boolean;
  progress10: boolean;
  progress11: boolean;
  progress12: boolean;
  progress13: boolean;
  progress14: boolean;
  progress15: boolean;
  progress16: boolean;
  progress17: boolean;
  progress18: boolean;
  progress19: boolean;
  progress20: boolean;
  progress21: boolean;
  progress22: boolean;
  progress23: boolean;
  progress24: boolean;
  progress25: boolean;
  progress26: boolean;
}


export { ReqProgressFormat};